% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{interpreter}
\alias{interpreter}
\title{Interpretation Function}
\usage{
interpreter(modelType = "none", df, dependentVar, independentVar,
  logDepen = F, logIndepen = NULL, squareIndepend = NULL,
  detail = FALSE)
}
\arguments{
\item{modelType}{What type of a regression model would you like to use? Defaults to "none"}

\item{df}{Define data frame.}

\item{dependentVar}{Insert the string value of the dependent variable you wish to use in this model.}

\item{independentVar}{Insert the vectors of string values of the independent variables you wish to use in this model.}

\item{logDepen}{Would you like to take the natural log of the dependent variable. Defaults to FALSE.}

\item{logIndepen}{Insert a string vector of independent variables that you would like to take the natural log of. Defaults to NULL.}

\item{squareIndepend}{Insert a string vector of independent variables that you would like to take the square of. Defaults to NULL.}

\item{detail}{Whether or not you would like to take the detailed output from the regressions.}
}
\description{
This function's main use is to run regressions and then returns interpretation of the model.
}
\examples{
interpreter()
}
\keyword{interpretation}
