% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizer}
\alias{optimizer}
\title{Find The Best Specifications}
\usage{
optimizer(data, depVar, indepVar = colnames(data)[colnames(data) != depVar],
  include = NULL, model = c("OLS", "binary probit", "binary logit",
  "ordered probit", "ordered logit", "multinomial logit"),
  time.series = FALSE, timeVar = NULL, sqIndep = NULL)
}
\arguments{
\item{data}{Data Frame}

\item{depVar}{Dependent Variable}

\item{indepVar}{a list of independent variables the subsets of which the dependent variable will be regressed on. Default is all variables in the data frame except for the dependent variable.}

\item{include}{a list of independent variables that the user wants to include in all regressions.}

\item{model}{regression model to estimate, including "OLS", "binary probit", "binary logit", "ordered probit", "ordered logit", "multinomial probit", "multinomial logit"}

\item{time.series}{Logical Variable.  This is used to decide the partition process. Defaults to FALSE.}

\item{timeVar}{Insert the time series variable. Defaults to NULL.}

\item{sqIndep}{Any independent variable that the user wants to consider quadratic terms for in the regressions. Defaults to NULL.}
}
\description{
This function searches for regression model specifications with the lowest MAPE, taking  the following inputs: a data frame, a dependent variable, a list of potential independent variables, and a regression model.
}
\examples{
##Use iris dataset
 @examples ##OLS model 
 @examples optimizer(iris, depVar = "Sepal.Length", include = "Sepal.Width", model = "OLS") 
Multinomial logistic model 
optimizer(iris, depVar = "Species", model = "multinomial logit")
}
\keyword{regressr}
