% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizer}
\alias{optimizer}
\title{Find The Best Specifications}
\usage{
optimizer(data, depVar, indepVar = colnames(data)[colnames(data) != depVar],
  include = NULL, model = c("OLS", "binary probit", "binary logit",
  "ordered probit", "ordered logit", "multinomial logit"),
  time.series = FALSE, timeVar = NULL, sqIndep = NULL)
}
\arguments{
\item{data}{a data frame}

\item{depVar}{a dependent variable}

\item{indepVar}{a list of independent variables the subsets of which the dependent variable will be regressed on. Defaul is all variables in the data frame but the dependent variable.}

\item{include}{a list of independent variables included in all regressions}

\item{model}{regression model to estimate, including "OLS", "binary probit", "binary logit", "ordered probit", "ordered logit", "multinomial probit", "multinomial logit"}

\item{time.series}{logical. This is used to decide the partition process.}

\item{timeVar}{the time series variable}

\item{sqIndep}{any variables in the independent variable list to consider quadratic terms in the regressions}
}
\description{
This function searches for regression model specifications with the lowest MAPE, taking inputs of a data frame, a dependent variable, a list of potential independent variables, and a regression model.
}
\examples{
##Use iris dataset 
##OLS model 
optimizer(iris, depVar = "Sepal.Length", include = "Sepal.Width", model = "OLS") 
##Multinomial logistic model 
optimizer(iris, depVar = "Species", model = "multinomial logit")
}
\keyword{regressr}
